refactor(core): narrow AuthorityIndex from u64 to u8 and cache rustup in Docker
Change AuthorityIndex type alias to u8 (max 128 authorities fits in
AuthoritySet's u128 bitmask) and propagate through committee, threshold
clock, cordial knowledge, and transaction generator. Add CryptoHash impl
for u8. Cache /usr/local/rustup in Dockerfile to avoid re-downloading
the toolchain on every build. Apply formatting fixes from nightly rustfmt.

+++ crates/starfish-core/src/transactions_generator.rs
@@ -39,7 +39,7 @@ impl TransactionGenerator {
         runtime::Handle::current().spawn(
             Self {
                 sender,
-                rng: StdRng::seed_from_u64(seed),
+                rng: StdRng::seed_from_u64(seed as u64),
                 parameters,
                 node_public_config,
                 metrics,
